bot/requirements.txt
python-telegram-bot>=21.4,<22
python-dotenv>=1.0,<2
google-api-python-client>=2.120,<3
google-auth>=2.28,<3
google-auth-httplib2>=0.2,<1
